# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: InnovateOS Klipper Installer API
  version: 1.0.0
  description: |
    API for the InnovateOS Klipper Installer, providing endpoints for board management,
    configuration, and installation of Klipper firmware.
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.innovateos.dev
    description: Production server
...
paths:
...
  /api/config/validate:
    post:
      summary: Validate printer configuration
      tags: [config]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrinterConfig'
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResult'
  /api/config/save:
    post:
      summary: Save printer configuration
      tags: [config]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrinterConfig'
      responses:
        '200':
          description: Configuration saved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigSaveResponse'
...
components:
  schemas:
...
    PrinterConfig:
      type: object
      required:
        - printer_name
        - kinematics
        - bed_size
        - board_type
        - mcu_path
      properties:
        printer_name:
          type: string
        kinematics:
          type: string
          enum: [cartesian, delta, corexy]
        bed_size:
          type: object
          properties:
            x: 
              type: number
            y:
              type: number
            z:
              type: number
        max_velocity:
          type: number
        max_accel:
          type: number
        board_type:
          type: string
        mcu_path:
          type: string
        features:
          type: object
          properties:
            pressure_advance:
              type: boolean
            input_shaping:
              type: boolean
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key